Refactor Command class into commands package
