+++ src/main/java/seedu/lifeasier/Parser.java
@@ -1,5 +1,7 @@
 package seedu.lifeasier;
 
+import seedu.lifeasier.commands.Command;
+
 public class Parser {
 
     public static Command parseCommand(String fullCommand) {
